# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/itemku/transactions/status": {
      "post": {
        "description": "Re-reads the supplier status for one reference id. NOTE: this route carries @UserRoles but no @UseGuards(AuthGuard), so no token is checked today and the role restriction is not enforced.",
        "operationId": "TransactionController_getPaymentTransactionStatus",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ItemkuRefIdDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Status of a payment transaction",
        "tags": [
          "Itemku Transaction"
        ]
      }
    },
...
    "/api/v1/tx/request-web": {
      "post": {
        "description": "Queues one transaction per entry in `customer_no` and returns the batch id. Requires BOTH a JWT and an `x-sign` header holding the HMAC-SHA256 hex digest of the request body. `username` is taken from the JWT, not from the body. Answers 200 with `errorCode` `4011` (missing signature) or `4012` (bad signature).",
        "operationId": "ExternalTransactionController_requestTransactionWeb",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExternalTxRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "x-sign": []
          },
          {
            "access-token": []
          }
        ],
        "summary": "Submit a transaction request (web session)",
        "tags": [
          "External Transaction"
        ]
      }
    },
    "/api/v1/tx/admin-balance": {
      "post": {
        "description": "Reads the balance of the account identified by the JWT.",
        "operationId": "ExternalTransactionController_getAdminBalance",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Balance of the calling admin",
        "tags": [
          "External Transaction"
        ]
      }
    },
    "/api/v1/tx/ext/admin-balance": {
      "post": {
        "description": "Machine-to-machine variant of `/admin-balance`: authenticate with the `x-username` + `x-sign` header pair instead of a JWT.",
        "operationId": "ExternalTransactionController_getAdminBalanceApi",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "x-sign": []
          },
          {
            "x-username": []
          }
        ],
        "summary": "Balance of the calling admin (API key)",
        "tags": [
          "External Transaction"
        ]
      }
    },
    "/api/v1/tx/status": {
      "post": {
        "description": "Re-reads the transaction status from the supplier by reference id.",
        "operationId": "ExternalTransactionController_getPaymentTxStatus",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefIdDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Status of a payment transaction",
        "tags": [
          "External Transaction"
        ]
      }
    },
    "/api/v1/tx/topup": {
      "post": {
        "description": "Sends a single top-up straight to the supplier, bypassing the batch queue. Unknown body properties are stripped by the validation pipe.",
        "operationId": "ExternalTransactionController_topUp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TopUpRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Direct supplier top-up",
        "tags": [
          "External Transaction"
        ]
      }
    },
    "/api/v1/tx/history": {
      "post": {
        "description": "Paginated list of transaction batches visible to the caller. Every filter is optional.",
        "operationId": "ExternalTransactionController_getTxHistory",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TxHistoryQueryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Transaction history grouped by batch",
        "tags": [
          "External Transaction"
        ]
      }
    },
    "/api/v1/tx/history-detail": {
      "post": {
        "description": "Returns every transaction belonging to the given `batch_id`.",
        "operationId": "ExternalTransactionController_getTxHistoryDetail",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchIdDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Transactions inside one batch",
        "tags": [
          "External Transaction"
        ]
      }
    },
    "/api/v1/tx/transaction-detail-list": {
      "post": {
        "description": "Same filters as `/history`, but returns individual transactions instead of batches.",
        "operationId": "ExternalTransactionController_getTxHistoryDetailByBatchId",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TxHistoryQueryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Flat, paginated transaction list",
        "tags": [
          "External Transaction"
        ]
      }
    },
    "/api/v1/tx/ext/request": {
      "post": {
        "description": "Machine-to-machine variant of `/request-web`. Authenticate with `x-username` + `x-sign`; the signature is verified against the caller API key over the raw body.",
        "operationId": "ExternalTransactionController_requestTransaction",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExternalTxRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "x-sign": []
          },
          {
            "x-username": []
          }
        ],
        "summary": "Submit a transaction request (API key)",
        "tags": [
          "External Transaction"
        ]
      }
    },
    "/api/v1/tx/ext/history-detail": {
      "post": {
        "description": "Machine-to-machine variant of `/history-detail`.",
        "operationId": "ExternalTransactionController_getTxHistoryDetailApi",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchIdDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "x-sign": []
          },
          {
            "x-username": []
          }
        ],
        "summary": "Transactions inside one batch (API key)",
        "tags": [
          "External Transaction"
        ]
      }
    },
    "/api/v1/tx/ext/tx-detail": {
      "post": {
        "description": "Looks up one transaction by its reference id.",
        "operationId": "ExternalTransactionController_getTxDetail",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefIdDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Single transaction detail",
        "tags": [
          "External Transaction"
        ]
      }
    },
    "/api/v1/tx/sign": {
      "post": {
        "description": "Helper endpoint: returns the HMAC-SHA256 hex digest of the posted JSON body, signed with the API key of the user named in the `x-username` header. Use the result as the `x-sign` header of the real request.",
        "operationId": "ExternalTransactionController_signPayload",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "The exact JSON body that will be sent to the signed endpoint.",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "example": {
                  "code": "ML5",
                  "customer_no": [
                    "12345678(1234)"
                  ]
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "example": "3b8c1f9e..."
                }
              }
            }
          }
        },
        "security": [
          {
            "x-username": []
          },
          {
            "access-token": []
          }
        ],
        "summary": "Sign a payload with a user API key",
        "tags": [
          "External Transaction"
        ]
      }
    },
    "/api/v1/tx/sign-sup": {
      "post": {
        "description": "Same as `/sign`, but signed with `PAYLOAD_SECRET`. This is the signature expected by `/request-web` and by the external top-up endpoints.",
        "operationId": "ExternalTransactionController_signPayloadSup",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "The exact JSON body that will be sent to the signed endpoint.",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "example": {
                  "amount": 100000,
                  "toAccount": "1234567890"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "example": "3b8c1f9e..."
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Sign a payload with the service secret",
        "tags": [
          "External Transaction"
        ]
      }
    },
...
  },
  "info": {
    "title": "TRX API Service",
    "description": "Transaction / top-up API service.\n\n## Authentication\n\nCall `POST /api/v1/auth/login` to obtain a JWT, then send it in the\n`authorization` header **without** the `Bearer ` prefix -- the guard verifies\nthe header value verbatim.\n\nMachine-to-machine endpoints (those under `/api/v1/tx/ext/...`) accept the\n`x-username` + `x-sign` header pair instead of a JWT. `x-sign` is the\nHMAC-SHA256 hex digest of the exact JSON request body signed with the\ncaller's API key (see `POST /api/v1/auth/api-key`).\n\nA few endpoints (`/api/v1/tx/request-web`, `/api/v1/external-topup/create-topup`,\n`/api/v1/external-topup/update-topup`) require BOTH a JWT and an `x-sign`\nheader over the body.\n\n## Response envelope\n\nMost endpoints answer with `{ message, data, errorCode }`. An `errorCode` of\n`\"0000\"` means success; any other code maps to an entry in\n`src/lib/errorCodes.ts` (e.g. `1000` wrong credential, `1002` insufficient\nbalance, `4012` signature not verified).",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "type": "apiKey",
        "in": "header",
        "name": "authorization",
        "description": "Raw JWT returned by POST /api/v1/auth/login. Do NOT prefix it with \"Bearer \"."
      },
      "x-username": {
        "type": "apiKey",
        "in": "header",
        "name": "x-username",
        "description": "Username of the API consumer. Used together with x-sign."
      },
      "x-sign": {
        "type": "apiKey",
        "in": "header",
        "name": "x-sign",
        "description": "HMAC-SHA256 hex signature of the JSON request body."
      }
    },
    "schemas": {
      "ItemkuRefIdDto": {
        "type": "object",
        "properties": {
          "ref_id": {
            "type": "string",
            "example": "TRX-20250901-0001",
            "description": "Transaction reference id"
          }
        },
        "required": [
          "ref_id"
        ]
      },
      "ApiResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "success",
            "description": "Human readable result message"
          },
          "data": {
            "type": "object",
            "nullable": true,
            "description": "Endpoint specific payload. `null` when the call failed."
          },
          "errorCode": {
            "type": "string",
            "example": "0000",
            "description": "`0000` on success, otherwise a code from src/lib/errorCodes.ts"
          }
        },
        "required": [
          "message",
          "data"
        ]
      },
...
      "ExternalTxRequestDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "example": "johndoe",
            "description": "Owner of the transaction. On `/request-web` it is overwritten with the JWT subject, so the value sent there is ignored."
          },
          "code": {
            "type": "string",
            "example": "ML5",
            "description": "Product code to purchase"
          },
          "customer_no": {
            "example": [
              "12345678(1234)",
              "87654321(4321)"
            ],
            "description": "Destination account(s). One transaction is queued per entry.",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "username",
          "code",
          "customer_no"
        ]
      },
      "RefIdDto": {
        "type": "object",
        "properties": {
          "ref_id": {
            "type": "string",
            "example": "TRX-20250901-0001",
            "description": "Reference id of a single transaction"
          }
        },
        "required": [
          "ref_id"
        ]
      },
      "TopUpRequestDto": {
        "type": "object",
        "properties": {
          "buyer_sku_code": {
            "type": "string",
            "example": "xld10",
            "description": "Supplier SKU code"
          },
          "customer_no": {
            "type": "string",
            "example": "087800001233",
            "description": "Destination account / phone number"
          },
          "testing": {
            "type": "boolean",
            "default": false,
            "description": "Run against the supplier sandbox"
          },
          "allow_dot": {
            "type": "boolean",
            "default": false,
            "description": "Allow decimal amounts"
          },
          "max_price": {
            "type": "number",
            "example": 100000,
            "description": "Reject the order above this price"
          }
        },
        "required": [
          "buyer_sku_code",
          "customer_no"
        ]
      },
      "TxHistoryQueryDto": {
        "type": "object",
        "properties": {
          "page": {
            "type": "number",
            "example": 1,
            "description": "Page number, 1 based"
          },
          "size": {
            "type": "number",
            "example": 10,
            "description": "Rows per page"
          },
          "customer_no": {
            "type": "string",
            "example": "087800001233",
            "description": "Filter by destination account"
          },
          "start_date": {
            "type": "string",
            "example": "2025-09-01",
            "description": "Inclusive lower bound (YYYY-MM-DD)"
          },
          "end_date": {
            "type": "string",
            "example": "2025-09-30",
            "description": "Inclusive upper bound (YYYY-MM-DD)"
          },
          "batch_id": {
            "type": "string",
            "description": "Filter by batch id"
          },
          "ref_id": {
            "type": "string",
            "description": "Filter by transaction reference id"
          }
        }
      },
      "BatchIdDto": {
        "type": "object",
        "properties": {
          "batch_id": {
            "type": "string",
            "example": "7c1d2e3f-4a5b-6c7d-8e9f-0a1b2c3d4e5f",
            "description": "Batch id returned when the transaction request was accepted"
          }
        },
        "required": [
          "batch_id"
        ]
      },
...
    }
  }
}